Add css attributes to html elements
Changes header elements to div elements and adds classes to elements of
body in preperation for styling. Headings all fall under header class
with a main or section class to differentiate each heading. Lists are
given list class as well as an identifying class for what each list
contains. Same convention will be applied accross all html files in
repo for consistency and ease of styling/cutting down on css.