Replace div with span in .prep
Tried to use <div> to bold only the beginning of each line of the list.
Found that <div> was causing the text following the element to move down
a line. Changed the <div> elements to <span> allowing me to bold the
beginning of each line using css without splitting each list element
into two lines.